requirements.txt
bcrypt==3.1.4
flask-restful==0.3.6
SQLAlchemy==1.2.0b3
object-mapper==1.0.0
Flask-Cors==3.0.3